ci: add manual trigger to bump the versions accordingly
This GH Actions workflow uses the previously created script to bump the versions automatically. The workflow can be triggered manually as a `workflow_dispatch`, and you just have to give it the input of wanting to do a patch release, a minor release, or a major release. It then bumps to the appropriate release version, commits it and tags it. Afterwards, it automatically bumps the version to the next snapshot version.

As the new release version is tagged, a release can be created when checking out this tag.